dj/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.8
configparser==6.0.0
Django==4.2.6
djangorestframework==3.14.0
etelemetry==0.3.1
filelock==3.12.4
future==0.18.3
httplib2==0.22.0
idna==3.4
isodate==0.6.1
kombu==5.3.2
looseversion==1.3.0
lxml==4.9.3
networkx==3.2
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.1
packaging==23.2
pandas==2.1.1
pathlib==1.0.1
Pillow==10.1.0
prompt-toolkit==3.0.39
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
redis==5.0.1
requests==2.31.0
scipy==1.11.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
traits==6.3.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
